Change the structure to make AntD modal aware of RFF
