+++ assets/src/application/ui/forms/formModal/FormModalForm.tsx
@@ -1,18 +1,13 @@
 import React, { useEffect, useState, CSSProperties } from 'react';
 import { __ } from '@wordpress/i18n';
+import { ButtonProps } from 'antd/lib/button';
+import { Modal } from 'antd';
 
 import { FormModalFormProps } from './types';
-import { EspressoButton, EspressoButtonType } from '../../input';
 
 const formStyle: CSSProperties = {
 	boxSizing: 'border-box',
 	padding: '1em 2em',
-};
-
-const btnRowStyle: CSSProperties = {
-	boxSizing: 'border-box',
-	padding: '1em 2em',
-	textAlign: 'right',
 };
 
 const dataStyle: CSSProperties = {
@@ -21,6 +16,20 @@ const dataStyle: CSSProperties = {
 	padding: '1em 2em',
 	color: '#a9ce47',
 	backgroundColor: '#26203d',
+};
+
+const modalStyle: CSSProperties = {
+	boxSizing: 'border-box',
+	maxHeight: '90%',
+	maxWidth: '1200px',
+	minHeight: '50%',
+	minWidth: '320px',
+	width: '80%',
+	overflowY: 'scroll',
+	position: 'absolute' as 'absolute',
+	left: '50%',
+	top: '50%',
+	transform: 'translate(-50%, -50%)',
 };
 
 const FormModalForm: React.FC<FormModalFormProps> = ({
@@ -45,42 +54,52 @@ const FormModalForm: React.FC<FormModalFormProps> = ({
 		}
 	});
 
+	const submitButton: ButtonProps = {
+		disabled: submitting || pristine,
+		htmlType: 'submit',
+		icon: 'save',
+		onClick: (click) => {
+			click.preventDefault();
+			form.submit();
+			onClose(click);
+		},
+	};
+
+	const resetButton: ButtonProps = {
+		disabled: submitting || pristine,
+		htmlType: 'reset',
+		onClick: (click) => {
+			click.preventDefault();
+			setFormReset(true);
+		},
+	};
+
 	return (
-		<form onSubmit={handleSubmit}>
-			<div style={formStyle}>
-				<FormComponent
-					form={form}
-					values={values}
-					submitting={submitting}
-					pristine={pristine}
-					formReset={formReset}
-					{...formProps}
-				/>
-			</div>
-			<div style={btnRowStyle}>
-				<EspressoButton
-					icon={'save'}
-					type={'submit'}
-					btnType={EspressoButtonType.PRIMARY}
-					buttonText={__('Submit')}
-					disabled={submitting || pristine}
-					onClick={(click) => {
-						click.preventDefault();
-						form.submit();
-						onClose(click);
-					}}
-				/>
-				<EspressoButton
-					buttonText={__('Reset')}
-					disabled={submitting || pristine}
-					onClick={(click) => {
-						click.preventDefault();
-						setFormReset(true);
-					}}
-				/>
-			</div>
-			<pre style={dataStyle}>{JSON.stringify(values, null, 2)}</pre>
-		</form>
+		<Modal
+			visible={true}
+			onOk={onClose}
+			style={modalStyle}
+			okText={__('Submit')}
+			okButtonProps={submitButton}
+			cancelText={__('Reset')}
+			cancelButtonProps={resetButton}
+			onCancel={onClose}
+			width={'80%'}
+		>
+			<form onSubmit={handleSubmit}>
+				<div style={formStyle}>
+					<FormComponent
+						form={form}
+						values={values}
+						submitting={submitting}
+						pristine={pristine}
+						formReset={formReset}
+						{...formProps}
+					/>
+				</div>
+				<pre style={dataStyle}>{JSON.stringify(values, null, 2)}</pre>
+			</form>
+		</Modal>
 	);
 };
